Call get points API call , closed #8

+++ app/src/main/java/com/example/mapboxexample/data/model/PointServer.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class PointServer(
     @SerializedName("id") val id: Long?,
-    @SerializedName("latitude") val latitude: Long?,
-    @SerializedName("longitude") val longitude: Long?
+    @SerializedName("latitude") val latitude: Float?,
+    @SerializedName("longitude") val longitude: Float?
 )
